+++ apps/omsorgspengerutbetaling-soknad/env.schema.ts
@@ -1,9 +1,7 @@
 import { commonEnvSchema } from '@navikt/sif-common-env';
 import * as z from 'zod';
 
-export const appEnvSchema = commonEnvSchema.extend({
-    SIF_PUBLIC_FEATURE_NYNORSK: z.enum(['on', 'off']),
-});
+export const appEnvSchema = commonEnvSchema;
 
 export const appDevEnvSchema = appEnvSchema.extend({
     MOCK_DATE: z.string().optional(),