+++ apps/omsorgspengerutbetaling-soknad/src/app/utils/appEnv.ts
@@ -1,8 +1,3 @@
-import { getCommonEnv, getMaybeEnv } from '@navikt/sif-common-env';
+import { getCommonEnv } from '@navikt/sif-common-env';
 
-const getAppEnv = () => ({
-    ...getCommonEnv(),
-    SIF_PUBLIC_FEATURE_NYNORSK: getMaybeEnv('SIF_PUBLIC_FEATURE_NYNORSK') === 'on' ? 'on' : 'off',
-});
-
-export const appEnv = getAppEnv();
+export const appEnv = getCommonEnv();