+++ apps/pleiepenger-i-livets-sluttfase-soknad/env.schema.ts
@@ -2,7 +2,6 @@ import { commonEnvSchema } from '@navikt/sif-common-env';
 import * as z from 'zod';
 
 export const appEnvSchema = commonEnvSchema.extend({
-    SIF_PUBLIC_FEATURE_NYNORSK: z.enum(['on', 'off']),
     SIF_PUBLIC_FEATURE_SOKE_TIDLIGERE: z.enum(['on', 'off']),
 });