+++ packages/sif-common-ui/src/inputs/velg-barn-input/i18n/nn.ts
@@ -1,8 +1,0 @@
-export const registrertBarnInput_nn = {
-    '@ui.registrertBarnInput.legend': 'Barn registrert på deg',
-    '@ui.registrertBarnInput.gjelderAnnetBarn': 'Søknaden gjelder et annet barn',
-    '@ui.registrertBarnInput.født': 'Født {dato}',
-    '@ui.registrertBarnInput.annetBarn.tittel': 'Annet barn',
-    '@ui.registrertBarnInput.kilde':
-        'Lista viser barn som er registrert på deg i Folkeregisteret. Viss dette ikkje stemmer, kan du kontakte Skatteetaten.',
-};