+++ packages/config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # config
 
+## 0.11.11
+
+### Patch Changes
+
+- Pakkeoppdateringer
+
+## 0.11.10
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/major + zod4
+
 ## 0.11.9
 
 ### Patch Changes