+++ packages/sif-common-hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/sif-common-hooks
 
+## 0.3.3
+
+### Patch Changes
+
+- Pakkeoppdateringer
+
 ## 0.3.2
 
 ### Patch Changes